# Conflicts:
#	designSystem/src/main/res/values-ar/strings.xml
#	presentation/src/main/java/com/madrid/presentation/screens/detailsScreen/seriesDetails/SeriesDetailsScreen.kt
#	presentation/src/main/java/com/madrid/presentation/screens/exceptionScreen/EmptySearchScreen.kt
#	presentation/src/main/java/com/madrid/presentation/screens/exceptionScreen/NoHistoryScreen.kt
#	presentation/src/main/java/com/madrid/presentation/screens/exceptionScreen/NoInternetScreen.kt
#	presentation/src/main/java/com/madrid/presentation/screens/exceptionScreen/NoSearchResultsScreen.kt
#	presentation/src/main/java/com/madrid/presentation/screens/exceptionScreen/exceptionsText.kt
#	presentation/src/main/res/values/strings.xml

+++ presentation/src/main/java/com/madrid/presentation/viewModel/detailsViewModel/SimilarMediaViewModel.kt
@@ -26,7 +26,7 @@ class SimilarMediaViewModel(
     private fun loadSimilarMedia() {
         tryToExecute(
             function = {
-                getSimilarMoviesUseCase(args.mediaId.toInt())
+                getSimilarMoviesUseCase(args.mediaId)
             },
             onSuccess = { allMovies ->
                 updateState {
@@ -46,7 +46,7 @@ class SimilarMediaViewModel(
     private fun loadSimilarSeries() {
         tryToExecute(
             function = {
-                getSimilarSeriesUseCase(args.mediaId.toInt())
+                getSimilarSeriesUseCase(args.mediaId)
             },
             onSuccess = { allSeries ->
                 updateState {
